separated trade to a different file and added trade_crypto() function in main loop
