+++ tickers/tickers_BTC.txt
@@ -1,273 +1,271 @@
-INJ
-ADA
-BQX
-ARK
-KNC
-JST
-CTSI
-GVT
-OM
-ARDR
-VITE
-GTO
-UNFI
-BAT
-DASH
-DODO
-SNM
-LSK
-VIB
-DATA
-KLAY
-TCT
-WPR
-XMR
-AKRO
-BNT
-FTM
-XTZ
-ENJ
-1INCH
-AXS
-DLT
-FIL
-FUN
-AGIX
-GRT
-MDA
-WABI
-REP
-CTXC
-CDT
-OG
-SYS
-LPT
-KMD
-XRP
-AUTO
-NEO
-SOL
-NAV
-CRV
-SFP
-REEF
-COS
-RIF
-THETA
-EOS
-ARPA
-NEAR
-BAL
-VIA
-PNT
-AERGO
+NKN
+SNT
+EZ
+BEAM
 TRX
-GO
-SAND
+BAL
 OMG
+KNC
+MITH
+FIRO
+NULS
+DNT
+PNT
+KEEP
+AUCTION
+BCD
+ASR
+XTZ
+SUSHI
+WTC
+SUSD
+HIVE
+DREP
+APPC
+OM
+LPT
+QSP
+BNT
+BTS
+CDT
+XVS
+IRIS
+MDA
+REQ
+HARD
+BQX
+NANO
+KLAY
+CELR
+PSG
+SKY
+YFII
+NAS
+SOL
+ICP
+REP
+RSR
+BAKE
+UMA
+LUNA
+YOYO
+NBS
+ORN
+HNT
+CHR
+OG
+DEGO
+LOOM
+CVC
+OGN
+AERGO
+POND
+XRP
+GRT
+ARDR
+AKRO
+POWR
+TVK
+AION
+KMD
+AR
+NU
+ONE
+SNX
+TWT
+AUDIO
+RLC
+NAV
+WAN
+VIDT
+STRAX
+ZIL
+CKB
+ICX
+PHA
+NEAR
+ATA
+PERP
+LRC
+RVN
+AAVE
+IDEX
+BEL
+OAX
+GTC
+ACM
+ANT
+TORN
+ALGO
+PHB
+ONT
+VITE
+FOR
+FET
+UNI
+POLY
+KSM
+XMR
+SCRT
+BAT
+ATOM
+BADGER
+QLC
+STX
+LSK
+REN
+KAVA
+AVA
+MKR
+RDN
+ADA
+MTH
+SAND
+QKC
+RCN
+FIO
+DODO
+TKO
+WING
+WABI
+FTT
+VIA
+NXS
+TOMO
+TFUEL
+AST
+JST
+SRM
+CAKE
+CHZ
+NMR
+NEBL
+GLM
+ETC
+QTUM
+DGB
+DUSK
+MANA
+IOTA
+FORTH
+CELO
+UTK
+GAS
+REEF
+COMP
+EGLD
+POLS
+IOTX
+ZEC
+ZRX
+ALICE
+POA
+COS
+GXS
 FLM
-XVG
+ETH
+RAMP
+RIF
+WAVES
+BRD
+MIR
+BAND
+BTG
+ENJ
+STMX
+SC
+SNM
+DASH
+BNB
+AGIX
+STEEM
+CTK
+ARK
+CTSI
+BCH
+WNXM
+CTXC
+LINA
+PIVX
+SXP
+XLM
+VIB
+LIT
+FTM
+DIA
+TLM
+ZEN
+GTO
+PPT
+PAXG
+FUN
+MDT
+DCR
+ANKR
+AXS
+ARPA
 IOST
 ALPHA
-CELR
-ONE
-CTK
-APPC
-PHA
-XLM
-BTG
-HNT
-NBS
-HBAR
-KAVA
-BZRX
-SUSHI
-ELF
-ANT
-LIT
-STRAX
-SUSD
-BRD
-JUV
-LINA
-RDN
-TRB
-BTS
-XEM
-CKB
-BAKE
-TKO
-ACM
-MTL
-HARD
-NKN
-MDT
-ROSE
-PSG
-RAMP
-NU
-MIR
-AMB
-FRONT
-PERP
-YOYO
-SRM
-CVC
-PPT
-KSM
-LRC
-STORJ
-ASR
-ZIL
-UNI
-ZRX
-SKL
-MANA
-CELO
-IRIS
-GRS
-DUSK
-IOTX
-HIVE
-ETC
-NANO
-ATM
-AAVE
-OXT
-POWR
-DOGE
-AVAX
-NEBL
-WTC
-STMX
-ZEN
-SNX
-DOCK
-WNXM
-GTC
-DGB
-MDX
-CAKE
-IDEX
-EGLD
-QSP
-DNT
-COMP
-KEEP
-CFX
-YFI
-ATOM
-STPT
-FIO
-STEEM
-NAS
-LUNA
-BEAM
-POND
-AUCTION
-TOMO
-WBTC
-ONG
-RLC
-SC
-LINK
-BCH
-ONT
-AION
-AST
-BEL
-ATA
-OGN
-PAXG
-DIA
-SNGLS
-BAND
-CHZ
-REN
-POA
-LTO
-BAR
+INJ
 WRX
-NULS
-GXS
-UMA
-MATIC
-MKR
-BADGER
-TRU
-ICX
-FOR
-QLC
+EOS
+1INCH
 EVX
-POLS
-MTH
-FORTH
-EZ
-TVK
-TWT
-IOTA
-ICP
-TFUEL
-SCRT
-PIVX
-AR
-TORN
-FTT
-GLM
-DEGO
-RVN
-ETH
-FIRO
-WAVES
-MITH
-CHR
-FET
-DREP
-BLZ
-RUNE
-ALICE
-SNT
-OCEAN
-TLM
-STX
 PERL
-BTCST
-EPS
-CND
-FIS
-GAS
-VIDT
-DCR
-SKY
-ORN
-COTI
-QKC
-NXS
-BNB
-POLY
-AVA
-XVS
-LOOM
-FXS
-RSR
-WING
-WAN
-RCN
-YFII
 ADX
-AUDIO
-VET
-ANKR
-SXP
-PHB
-OAX
-LTC
-BCD
-NMR
-ALGO
-TROY
+CRV
+MATIC
+NEO
+STPT
+AVAX
+ATM
 DOT
-UTK
-QTUM
-REQ
-ZEC
+TCT
+SKL
+DOGE
+STORJ
+XVG
+OCEAN
+MDX
+VET
+OXT
+WPR
+HBAR
+DATA
+RUNE
+LTC
+SYS
+JUV
+CFX
+ONG
+FXS
+BLZ
+AUTO
+XEM
+ROSE
+LTO
+CND
+ELF
+TROY
+GRS
+FIL
+TRU
+FRONT
+AMB
+BAR
+SNGLS
+EPS
+YFI
+COTI
+GO
+DLT
+TRB
+THETA
+LINK
+FIS
+BZRX
+DOCK
+SFP
+MTL
+UNFI
+GVT
